# ARROW-10968: [Rust][DataFusion] Don't build hash table for right side of join